backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.30
asyncpg>=0.29.0
pydantic>=2.7.0
bcrypt>=4.1.0
PyJWT>=2.8.0
